Move non-command CLI parsers from Cardano.CLI.Legacy.Options to Cardano.CLI.EraBased.Options.Common
